Refactor Inventory Screen.
Tidy up the messy code into seperate components that share the right data.
